fix(desktop): unblock the specta TypeScript export (BigIntForbidden)
Every `tauri dev` since 3d7cc9417a (2026-08-14) panicked at startup:

  panicked at src/lib.rs:702: Failed to export typescript bindings: BigIntForbidden

Root cause: WorkbenchLease.issued_at/expires_at and
WorkbenchRotation.grace_period_ms were declared u64.
specta_typescript::Typescript::default() uses BigIntExportBehavior::Fail,
so export_types() panics on any 64-bit integer. Release builds never hit
it because export_types is #[cfg(any(debug_assertions, test))] — which is
why `tauri build` worked while `tauri dev` was dead for four days.

Fix follows the convention this codebase already documents three times in
llm.rs (ModelInfo.size, VramInfo, BenchmarkResult): use f64. Epoch
milliseconds are exact in f64 up to 2^53 (year 285616), and the
TypeScript side already types these fields as `number` and validates them
with Number.isSafeInteger — see workbench-shell/src/native-token-bridge.ts.

Evidence:
  - bindings.ts was last regenerated 2026-05-07 and contained zero
    Workbench types or commands, confirming the export had been failing
    since the bridge landed.
  - After the fix the debug binary boots past setup and regenerates
    bindings.ts (4070 -> 9845 bytes), now exporting WorkbenchLease,
    WorkbenchRotation and the four workbench_* commands.
  - cargo check: clean. Project-wide scan for the same pattern across
    desktop, mobile and crates: no remaining 64-bit field in a
    specta::Type.

Also excludes the generated bindings.ts from biome, mirroring the
existing "!packages/sdk/js/src/**/gen" rule. The file declares "Do not
edit this file manually"; linting it only produced noise that invites
--no-verify.

+++ packages/desktop/src-tauri/src/lib.rs
@@ -94,8 +94,12 @@ struct WorkbenchLease {
     instance_id: String,
     workspace_id: String,
     capabilities: Vec<String>,
-    issued_at: u64,
-    expires_at: u64,
+    // f64: specta rejects u64 (BigIntForbidden); epoch milliseconds are exact
+    // in f64 up to 2^53 (year 285616). The TypeScript side already types these
+    // as `number` and validates them with Number.isSafeInteger — see
+    // packages/workbench-shell/src/native-token-bridge.ts.
+    issued_at: f64,
+    expires_at: f64,
 }
 
 #[derive(Clone, serde::Serialize, serde::Deserialize, specta::Type)]
@@ -103,7 +107,9 @@ struct WorkbenchLease {
 struct WorkbenchRotation {
     token: WorkbenchLease,
     previous_token: Option<String>,
-    grace_period_ms: u64,
+    // f64: specta rejects u64 (BigIntForbidden); a grace period in milliseconds
+    // is orders of magnitude below f64's exact-integer range.
+    grace_period_ms: f64,
 }
 
 async fn workbench_native_request(
